extract and refactor createPortalSlot
